Defined a Parser type for parsing in const contexts.
For now I've only defined methods for matching on the start and end of strings,
and getting the unmatched portion of it.

Defined an `assign_if` macro to emulate calling `&mut self` methods on the Parser,
but it can be used for other stuff too.

Added "parsing" feature to enable the parser.