Added doc examples of parsing structs in root module and `rebind_if_ok` macro.
Removed `offset_end` field from Parser, since it's redundant.

Added `parsing_direction: ParseDirection` field to parser so that
it knows in what direction it was parsing.

Removed the `ParseDirection` parameter of the `Parser::into*error` methods
and `ParserError::new`, now it's taken from the `Parser` itself.

Defined `for_range` macro to do `for pattern in start..end` in const contexts