Fixed string::RSplitTerminator behavior
`RSplitTerminator` used to skip the empty string before the first instance of
the delimiter, instead of the empty space after the last one.
The latter is what the iterator does now, to be consistent with std.

Removed `,rev()` method of `konst::string::{Split, RSplit}`, because it allowed reversing an iterator with a `&str` needle, when std does not allow it (because the reverse iterator can produce different strings than forward).

Added (ui) test for failed conversions from slice to array

Updated tests for splitting strings, added some cases,
added crosschecking with `str::split_*` methods.